+++ test/converters.test.ts
@@ -730,3 +730,41 @@ test("literal string converter", () => {
     "kwarktaart"
   );
 });
+
+test("zero decimal empty and required", () => {
+  const M = types.model("M", {
+    foo: types.string,
+  });
+
+  const form = new Form(M, {
+    foo: new Field(
+      converters.stringDecimal({ decimalPlaces: 2, zeroIsEmpty: true }),
+      { required: true }
+    ),
+  });
+
+  const o = M.create({ foo: "0" });
+
+  const state = form.state(o);
+  const field = state.field("foo");
+
+  // We expect the field to be empty and required, since zeroIsEmpty is true and the raw is "0".
+  expect(field.isEmptyAndRequired).toBeTruthy();
+
+  // We expect the field to still be empty and required, since zeroIsEmpty is true and the raw is "0.00".
+  // This is equivalent to 0.
+  field.setRaw("0.00");
+
+  expect(field.isEmptyAndRequired).toBeTruthy();
+
+  // We expect the field to still be empty and required, since zeroIsEmpty is true and the raw is "".
+  // This is equivalent to the empty raw.
+  field.setRaw("");
+
+  expect(field.isEmptyAndRequired).toBeTruthy();
+
+  // We no longer expect the field to be empty and required, since zeroIsEmpty is true but the raw is "0.0012".
+  // This is equivalent to 0.0012, which isn't technically 0.
+  field.setRaw("0.0012");
+  expect(field.isEmptyAndRequired).toBeFalsy();
+});